+++ src/main/java/lotto/LottoFactory.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.Lotto;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
