+++ src/main/java/lotto/domain/WinningRank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Collection;
 import java.util.HashMap;
